[PDS-168] refactor: 예약 관련 서비스 분리
